(AKA-64) Вынес функции в отдельный файл
