+++ Map/Scenes/Smart target list scene/Detail target scene/DetailTargetViewController.swift
@@ -414,85 +414,6 @@ extension DetailTargetViewController
 			self.activityIndicator.centerYAnchor.constraint(equalTo: addressLabel.centerYAnchor),
 		])
 	}
-
-	private func setupAnnotation() {
-		let annotaion = presenter.getAnnotation()
-		mapView.addAnnotation(annotaion)
-		mapView.showAnnotations([annotaion], animated: false)
-	}
-
-	private func setupOverlay() {
-		let overlay = presenter.getCircleOverlay()
-		mapView.addOverlay(overlay)
-	}
-
-	private func setSmartTargetRegion(coordinate: CLLocationCoordinate2D,
-									  region: MKCoordinateRegion? = nil,
-									  animated: Bool) {
-		let radius = presenter.editRadius * 4
-		var region = region ?? MKCoordinateRegion(center: coordinate, latitudinalMeters: radius, longitudinalMeters: radius)
-		region.center = coordinate
-		mapView.setRegion(region, animated: animated)
-	}
-}
-
-extension DetailTargetViewController: MKMapViewDelegate
-{
-	func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
-		if annotation is MKUserLocation { return nil }
-		let pinView = MKPinAnnotationView(annotation: annotation, reuseIdentifier: nil)
-		pinView.animatesDrop = false
-		pinView.isDraggable = true
-		pinView.canShowCallout = false
-		return pinView
-	}
-
-	func mapView(_ mapView: MKMapView, rendererFor overlay: MKOverlay) -> MKOverlayRenderer {
-		let renderer = MKCircleRenderer(overlay: overlay)
-		if #available(iOS 13.0, *) {
-			renderer.fillColor = UIColor.systemBackground.withAlphaComponent(0.5)
-		}
-		else {
-			renderer.fillColor = UIColor.white.withAlphaComponent(0.5)
-		}
-		renderer.strokeColor = .systemBlue
-		renderer.lineWidth = 1
-		return renderer
-	}
-
-	func mapView(_ mapView: MKMapView, didAdd renderers: [MKOverlayRenderer]) {
-	}
-
-	func mapView(_ mapView: MKMapView,
-				 annotationView view: MKAnnotationView,
-				 didChange newState: MKAnnotationView.DragState,
-				 fromOldState oldState: MKAnnotationView.DragState) {
-		switch (oldState, newState) {
-		case (.none, .starting): // 0 - 1
-			mapView.removeOverlays(mapView.overlays)
-		case (.starting, .dragging): // 1 - 2
-			addressLabel.text = nil
-			activityIndicator.startAnimating()
-		case (.dragging, .ending), // 2 - 4
-			 (.dragging, .canceling), // 2 - 3
-			 (.starting, .canceling), // 1 - 3
-			 (.starting, .ending): // 1 - 4
-			impactFeedbackGenerator.prepare()
-			presenter.editCoordinate = view.annotation?.coordinate ?? presenter.editCoordinate
-		case (.canceling, .none): // 3 - 0
-			impactFeedbackGenerator.impactOccurred()
-			setupOverlay()
-		case (.ending, .none): // 4 - 0
-			impactFeedbackGenerator.impactOccurred()
-			setSmartTargetRegion(coordinate: presenter.editCoordinate, region: mapView.region, animated: true)
-			presenter.getAddressText {
-				self.addressLabel.text = $0
-				self.activityIndicator.stopAnimating()
-			}
-			setupOverlay()
-		default: break
-		}
-	}
 }
 
 // MARK: - Actions
